flush拆分
